+++ plugins/chat/__init__.py
@@ -1,0 +1,1 @@
+__doc__ = """Pack with plugin for chats."""