+++ plugins/content/__init__.py
@@ -1,0 +1,1 @@
+__doc__ = """Pack with plugins that can generate content or information."""