+++ plugins/utils/__init__.py
@@ -1,0 +1,1 @@
+__doc__ = """Pack with plugins that are useful for other plugins."""